requirements.txt
anyio==3.7.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
decorator==5.1.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
pycparser==2.21
python-telegram-bot==12.8
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
tornado==6.3.2
urllib3==2.0.3
